requirements.txt
# ========================================================================
# PHOTO SOFTWARE - PYTHON DEPENDENCIES
# ========================================================================

# ----------------------------------------
# Web Framework
# ----------------------------------------
setuptools
fastapi
uvicorn
python-multipart
jinja2

# ----------------------------------------
# Gesichtserkennung
# ----------------------------------------
face-recognition
face-recognition-models
dlib

# ----------------------------------------
# Computer Vision
# ----------------------------------------
opencv-python
opencv-contrib-python
Pillow

# ----------------------------------------
# YOLO / Object Detection
# ----------------------------------------
ultralytics
torch
torchvision

# ----------------------------------------
# Machine Learning / Data Science
# ----------------------------------------
numpy
scikit-learn
scipy

# ----------------------------------------
# Dateisystem-Überwachung
# ----------------------------------------
watchdog

# ----------------------------------------
# Bildverarbeitung
# ----------------------------------------
imageio
scikit-image

# ----------------------------------------
# Utilities
# ----------------------------------------
python-dateutil
pytz

# ----------------------------------------
# Windows-spezifisch (Drucker)
# ----------------------------------------
pywin32

# ----------------------------------------
# Logging & Debugging
# ----------------------------------------
colorlog
python-json-logger

# ----------------------------------------
# Optional: Performance
# ----------------------------------------
numba 
orjson

# ----------------------------------------
# Development (optional)
# ----------------------------------------
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1